+++ app/models/measurement.rb
@@ -1,4 +1,4 @@
 class Measurement < ApplicationRecord
-    belongs_to :activity
-    validates_presence_of :duration, :date
+  belongs_to :activity
+  validates_presence_of :duration, :date
 end
