+++ app/models/user.rb
@@ -1,4 +1,4 @@
 class User < ApplicationRecord
-    has_secure_password
-    validates_presence_of :username, :password_digest, :age
+  has_secure_password
+  validates_presence_of :username, :password_digest, :age
 end
